Reduce unnecessary runtime dependencies
As far as I can make out, @types/chokidar, @type/node-sass,
jest and typescript are only needed for compiling this
package, not running it, so I've moved them into devDependencies.

The legacy 'tsc' package is redundant, so I've removed it:
'typescript' provides the 'tsc' binary needed for the compilation
step, not the 'tsc' package.